third_party/WebKit/LayoutTests/inspector/sources/debugger/custom-element-lifecycle-events.html

<html>
<head>
<script src="../../../http/tests/inspector/inspector-test.js"></script>
<script src="../../../http/tests/inspector/console-test.js"></script>
<script src="../../../http/tests/inspector/debugger-test.js"></script>
<script>

function test() {
    var setup = [
        "var proto = Object.create(HTMLElement.prototype);",
        "proto.createdCallback = function() { output('Invoked createdCallback.'); };",
        "proto.attachedCallback = function() { output('Invoked attachedCallback.'); };",
        "proto.detachedCallback = function() { output('Invoked detachedCallback.'); };",
        "proto.attributeChangedCallback = function() { output('Invoked attributeChangedCallback.'); };",
        "CustomElement = document.registerElement('x-foo', {prototype: proto});",
    ].join('\n');

    var lifecycleCallbacks = [
        "created = new CustomElement();",
        "created.setAttribute('x', '1');",
        "document.body.appendChild(created);",
        "created.remove();",
    ].join('\n');

    InspectorTest.startDebuggerTest(step1);

    function step1() {
        InspectorTest.evaluateInConsole(setup, function() {
            InspectorTest.addResult("Custom element registered.");
            InspectorTest.evaluateInConsoleAndDump("new CustomElement() instanceof CustomElement", step2);
        });
    }

    function step2() {
        InspectorTest.evaluateInConsole('debugger;');
        InspectorTest.waitUntilPaused(step3);
    }

    function step3() {
        InspectorTest.evaluateInConsoleAndDump("new CustomElement() instanceof CustomElement", step4);
    }

    function step4() {
        InspectorTest.evaluateInConsole(lifecycleCallbacks, InspectorTest.completeDebuggerTest);
    }
}

</script>
</head>

<body onload="runTest()">
<p>
Tests that custom element lifecycle events fire while debugger is paused.
</p>

</body>
</html>
