third_party/WebKit/LayoutTests/inspector/sources/debugger/custom-element-lifecycle-events-expected.txt

Tests that custom element lifecycle events fire while debugger is paused.

[page] Invoked createdCallback.
[page] Invoked createdCallback.
[page] Invoked createdCallback.
[page] Invoked attributeChangedCallback.
[page] Invoked attachedCallback.
[page] Invoked detachedCallback.
Custom element registered.
new CustomElement() instanceof CustomElement = true
Script execution paused.
new CustomElement() instanceof CustomElement = true
Script execution resumed.
